Cart and Compare

+++ database/migrations/2025_03_15_174419_create_reviews_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('reviews', function (Blueprint $table) {
             $table->id();
-            $table->double('rate');
+            $table->unsignedInteger('rate');
             $table->string('content')->nullable();
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('product_id');
